refactor: プロジェクト全体の構造をリファクタリングし、モジュール性を向上
目的：
将来的な機能拡張やメンテナンスを容易にするため、アプリケーションの構造を全面的に見直し、ドメイン（関心事）の分離を徹底する。

主な変更点：

1. **ディレクトリ構造の変更:**
   - `src`配下に`core`（共通基盤）と`modules`（機能別モジュール）を導入。
   - 従来の`scheduler`関連のファイルを`src/modules/scheduler`に集約。

2. **パッケージング設定の改善:**
   - `pyproject.toml`を更新し、`src`レイアウトに最適化されたモダンなパッケージング構成に変更。
   - これにより、Pythonのインポート解決がより堅牢になり、`PYTHONPATH`への依存を排除。

3. **インポートパスの全面的な修正:**
   - プロジェクト内のすべてのPythonファイルで、`src`をルートとした絶対インポートに統一。
   - これにより、`ModuleNotFoundError`や相対インポートに起因する問題を解消。

4. **APIエンドポイントの修正:**
   - フロントエンド(JavaScript)から呼び出されるすべてのAPIパスを、リファクタリング後の`/api/...`形式に更新。

5. **キャッシュと古いデータのクリーンアップ:**
   - リファクタリングに際し、古いバイトコードキャッシュ(`__pycache__`)、`egg-info`、`jobs.sqlite`データベースをすべて削除し、クリーンな状態から再構築。
