+++ jobs.yaml
@@ -1,6 +1,6 @@
 # jobs.yaml
 - id: 'daily_sales_report'
-  func: 'scheduler.tasks.reporting.send_daily_report'
+  func: 'modules.scheduler.tasks.reporting.send_daily_report'
   description: 'Generates and sends the daily sales report.'
   is_enabled: true
   trigger:
@@ -19,7 +19,7 @@
   misfire_grace_time: 3600 # 1 hour
 
 - id: 'api_health_check'
-  func: 'scheduler.tasks.monitoring.check_api_status'
+  func: 'modules.scheduler.tasks.monitoring.check_api_status'
   description: 'Checks the health of the main API endpoint every 5 minutes.'
   is_enabled: true
   trigger:
@@ -33,7 +33,7 @@
   coalesce: true
 
 - id: 'sample_job_1'
-  func: 'scheduler.tasks.sample_tasks.print_current_time'
+  func: 'modules.scheduler.tasks.sample_tasks.print_current_time'
   description: 'A sample job that prints the current time every 10 minutes.'
   is_enabled: true
   trigger:
@@ -42,7 +42,7 @@
   replace_existing: true
 
 - id: 'sample_job_2_paused'
-  func: 'scheduler.tasks.sample_tasks.print_current_time'
+  func: 'modules.scheduler.tasks.sample_tasks.print_current_time'
   description: 'A sample job that is initially disabled (paused).'
   is_enabled: false
   trigger:
