+++ src/core/config.py
@@ -1,10 +1,9 @@
 from pydantic_settings import BaseSettings, SettingsConfigDict
 
-class Settings(BaseSettings):
-    """Application settings"""
+class CoreSettings(BaseSettings):
+    """Core application settings"""
     DATABASE_URL: str = "sqlite:///jobs.sqlite"
 
     model_config = SettingsConfigDict(env_file=".env")
 
-
-settings = Settings()
+settings = CoreSettings()
