+++ src/modules/scheduler/tasks/backup.py
@@ -1,6 +1,7 @@
-from util import logger
+from util import logger_util
+
+logger = logger_util.get_logger(__name__)
 
 def backup_data():
     logger.info("Performing data backup...")
-    # Add your backup logic here
     pass
