+++ src/modules/scheduler/tasks/reporting.py
@@ -1,0 +1,4 @@
+import logging
+
+def send_daily_report(email: str, **kwargs):
+    logging.info(f"Sending daily report to {email}.")
