--- src/scheduler/main.py
@@ -1,709 +1,0 @@
-from contextlib import asynccontextmanager
-from typing import Generator, List, Optional
-from datetime import datetime, timedelta, timezone
-
-from fastapi import Depends, FastAPI, HTTPException, status
-from fastapi.middleware.cors import CORSMiddleware
-from apscheduler.jobstores.base import JobLookupError
-from sqlalchemy.orm import Session
-from pydantic import BaseModel
-
-# --- App Imports ---
-from scheduler import database
-from .models import JobDefinition, JobConfig, ErrorResponse, ProcessExecutionLog, ProcessExecutionLogInfo
-from .scheduler import scheduler, start_scheduler, shutdown_scheduler
-from .loader import load_and_validate_jobs, apply_job_config, start_config_watcher, sync_jobs_from_db, seed_db_from_yaml
-from util import logger
-
-
-@asynccontextmanager
-async def lifespan(app: FastAPI):
-    # --- Startup ---
-    logger.info("Application startup...")
-    config_path = "jobs.yaml"
-    
-    # 1. Initialize Database
-    database.init_db()
-
-    # 2. Seed the database from the YAML file to ensure it's up-to-date
-    logger.info(f"Seeding database from {config_path}...")
-    seed_db_from_yaml(config_path)
-    
-    # 3. Start Scheduler
-    start_scheduler()
-
-    # 4. Perform initial sync from DB to scheduler memory
-    logger.info("Performing initial job sync from database to scheduler...")
-    try:
-        sync_jobs_from_db()
-    except Exception as e:
-        logger.critical(f"Initial job sync failed: {e}", exc_info=True)
-
-    # 5. Start file watcher to hot-reload YAML changes
-    logger.info(f"Starting file watcher for {config_path}...")
-    watcher = start_config_watcher(scheduler, config_path)
-    
-    # 6. Schedule periodic DB sync
-    scheduler.add_job(
-        sync_jobs_from_db,
-        "interval",
-        seconds=60,
-        id="internal_db_sync",
-        replace_existing=True,
-    )
-    logger.info("Scheduled periodic job sync every 60 seconds.")
-
-    yield # Application is running
-
-    # --- Shutdown ---
-    logger.info("Application shutdown...")
-    if watcher:
-        watcher.stop()
-        watcher.join()
-        logger.info("File watcher stopped.")
-    shutdown_scheduler()
-
-app = FastAPI(title="Resilient Task Scheduler API", lifespan=lifespan)
-
-# CORS Middleware
-origins = [
-    "http://localhost:5012",
-    "http://127.0.0.1:5012",
-]
-
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=origins,
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
-)
-
-
-
-
-class TimelineItem(BaseModel):
-    id: str
-    content: str
-    start: datetime
-    end: Optional[datetime] = None
-    status: str # e.g., 'scheduled', 'running', 'completed', 'failed'
-    group: Optional[str] = None # For grouping in vis.js
-
-class JobInfo(JobConfig):
-    next_run_time: Optional[datetime] = None
-    # Add other relevant fields from APScheduler Job object if needed
-
-class BulkJobUpdate(BaseModel):
-    job_ids: List[str]
-
-# Dependency to get DB session
-def get_db() -> Generator[Session, None, None]:
-    if database.SessionLocal is None:
-        raise RuntimeError("Database is not initialized. SessionLocal is None.")
-    db = database.SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
-
-@app.get("/", tags=["Root"])
-def read_root():
-    return {"message": "Welcome to the Resilient Task Scheduler API!"}
-
-
-# --- Dashboard Endpoints ---
-
-class DashboardSummary(BaseModel):
-    total_jobs: int
-    running_jobs: int
-    successful_runs: int
-    failed_runs: int
-
-@app.get("/api/dashboard/summary", response_model=DashboardSummary, tags=["Dashboard"])
-def get_dashboard_summary(db: Session = Depends(get_db)):
-    """
-    Provides a summary of job statuses for the dashboard.
-    """
-    try:
-        total_jobs = db.query(JobDefinition).count()
-        
-        # Note: This is a simplified aggregation. For large datasets, more efficient queries would be needed.
-        running_jobs = db.query(ProcessExecutionLog).filter(ProcessExecutionLog.status == 'RUNNING').count()
-        successful_runs = db.query(ProcessExecutionLog).filter(ProcessExecutionLog.status == 'COMPLETED').count()
-        failed_runs = db.query(ProcessExecutionLog).filter(ProcessExecutionLog.status == 'FAILED').count()
-
-        return DashboardSummary(
-            total_jobs=total_jobs,
-            running_jobs=running_jobs,
-            successful_runs=successful_runs,
-            failed_runs=failed_runs,
-        )
-    except Exception as e:
-        logger.error(f"Error fetching dashboard summary: {e}", exc_info=True)
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Failed to fetch dashboard summary data."
-        )
-
-@app.get("/api/logs", response_model=List[ProcessExecutionLogInfo], tags=["Dashboard"])
-def get_execution_logs(skip: int = 0, limit: int = 100, db: Session = Depends(get_db)):
-    """
-    Retrieves a list of process execution logs from the database.
-    """
-    try:
-        logs = (
-            db.query(ProcessExecutionLog)
-            .order_by(ProcessExecutionLog.start_time.desc())
-            .offset(skip)
-            .limit(limit)
-            .all()
-        )
-        return logs
-    except Exception as e:
-        logger.error(f"Error fetching execution logs: {e}", exc_info=True)
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Failed to fetch execution logs."
-        )
-
-@app.get("/api/timeline/data", response_model=List[TimelineItem], tags=["Dashboard"])
-def get_timeline_data(db: Session = Depends(get_db)):
-    """
-    Provides data for the job execution timeline, including scheduled jobs and historical runs.
-    """
-    timeline_items: List[TimelineItem] = []
-
-    # 1. Add scheduled jobs from APScheduler
-    scheduled_jobs = scheduler.get_jobs()
-    for job in scheduled_jobs:
-        if job.next_run_time:
-            # Ensure next_run_time is timezone-aware (APScheduler usually provides this)
-            # If it's naive, assume UTC and make it aware.
-            start_time_aware = job.next_run_time
-            if start_time_aware.tzinfo is None:
-                start_time_aware = start_time_aware.replace(tzinfo=timezone.utc)
-
-            timeline_items.append(
-                TimelineItem(
-                    id=f"scheduled-{job.id}-{start_time_aware.isoformat()}",
-                    content=f"{job.id} (Scheduled)",
-                    start=start_time_aware,
-                    status="scheduled",
-                    group=job.id # Group by job ID
-                )
-            )
-        # Also add currently running jobs if APScheduler provides a way to identify them
-        # APScheduler doesn't directly expose 'running' status for jobs in get_jobs()
-        # This would typically come from ProcessExecutionLog where status is 'RUNNING'
-
-    # 2. Add historical job runs from the database
-    # Fetch recent logs, e.g., last 7 days or a reasonable limit
-    # For simplicity, let's fetch a recent batch. Adjust limit as needed.
-    recent_logs = (
-        db.query(ProcessExecutionLog)
-        .filter(ProcessExecutionLog.start_time >= datetime.now(timezone.utc) - timedelta(days=7)) # Last 7 days
-        .order_by(ProcessExecutionLog.start_time.asc())
-        .all()
-    )
-
-    for log in recent_logs:
-        item_status = log.status.lower() # COMPLETED -> completed, FAILED -> failed, RUNNING -> running
-        
-        # Ensure log datetimes are timezone-aware (assume UTC if naive from DB)
-        log_start_time_aware = log.start_time
-        if log_start_time_aware.tzinfo is None:
-            log_start_time_aware = log_start_time_aware.replace(tzinfo=timezone.utc)
-
-        log_end_time_aware = None
-        if log.end_time:
-            log_end_time_aware = log.end_time
-            if log_end_time_aware.tzinfo is None:
-                log_end_time_aware = log_end_time_aware.replace(tzinfo=timezone.utc)
-        
-        item_end = log_end_time_aware if log_end_time_aware else (datetime.now(timezone.utc) if item_status == 'running' else None)
-
-        timeline_items.append(
-            TimelineItem(
-                id=f"log-{log.id}",
-                content=f"{log.job_id} ({item_status.capitalize()})",
-                start=log_start_time_aware,
-                end=item_end,
-                status=item_status,
-                group=log.job_id # Group by job ID
-            )
-        )
-    
-    # Sort items by start time for better visualization
-    timeline_items.sort(key=lambda item: item.start)
-
-    return timeline_items
-
-
-# --- Job Definition Management Endpoints (Database) ---
-
-@app.get("/jobs", response_model=List[JobConfig], tags=["Job Definitions"])
-def read_jobs(db: Session = Depends(get_db)):
-    jobs = db.query(JobDefinition).all()
-    return [JobConfig.model_validate(job) for job in jobs]
-
-@app.post(
-    "/jobs",
-    response_model=JobConfig,
-    status_code=status.HTTP_201_CREATED,
-    tags=["Job Definitions"],
-    responses={
-        status.HTTP_409_CONFLICT: {"model": ErrorResponse, "description": "Conflict: Job with this ID already exists"}
-    }
-)
-def create_job(job: JobConfig, db: Session = Depends(get_db)):
-    db_job = db.query(JobDefinition).filter(JobDefinition.id == job.id).first()
-    if db_job:
-        logger.warning(f"Attempted to create job with existing ID: {job.id}")
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail="Job with this ID already exists"
-        )
-    
-    # Map JobConfig Pydantic model to JobDefinition SQLAlchemy model
-    trigger_dict = job.trigger.dict()
-    trigger_type = trigger_dict.pop('type')
-    trigger_config = trigger_dict
-
-    db_job = JobDefinition(
-        id=job.id,
-        func=job.func,
-        trigger_type=trigger_type,
-        trigger_config=trigger_config,
-        args=job.args,
-        kwargs=job.kwargs,
-        max_instances=job.max_instances,
-        coalesce=job.coalesce,
-        misfire_grace_time=job.misfire_grace_time,
-    )
-    db.add(db_job)
-    db.commit()
-    db.refresh(db_job)
-    logger.info(f"Job '{job.id}' created successfully.")
-    sync_jobs_from_db() # Sync APScheduler with the new database state
-    return JobConfig.model_validate(db_job)
-
-@app.get(
-    "/jobs/{job_id}",
-    response_model=JobConfig,
-    tags=["Job Definitions"],
-    responses={
-        status.HTTP_404_NOT_FOUND: {"model": ErrorResponse, "description": "Not Found: Job not found"}
-    }
-)
-def read_job(job_id: str, db: Session = Depends(get_db)):
-    job = db.query(JobDefinition).filter(JobDefinition.id == job_id).first()
-    if job is None:
-        logger.warning(f"Attempted to read non-existent job: {job_id}")
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Job not found"
-        )
-    logger.info(f"Job '{job_id}' read successfully.")
-    return JobConfig.model_validate(job)
-
-@app.put(
-    "/jobs/{job_id}",
-    response_model=JobConfig,
-    tags=["Job Definitions"],
-    responses={
-        status.HTTP_404_NOT_FOUND: {"model": ErrorResponse, "description": "Not Found: Job not found"}
-    }
-)
-def update_job(job_id: str, job: JobConfig, db: Session = Depends(get_db)):
-    db_job = db.query(JobDefinition).filter(JobDefinition.id == job_id).first()
-    if db_job is None:
-        logger.warning(f"Attempted to update non-existent job: {job_id}")
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Job not found"
-        )
-    
-    # Update fields
-    db_job.func = job.func
-    
-    trigger_dict = job.trigger.dict()
-    db_job.trigger_type = trigger_dict.pop('type')
-    db_job.trigger_config = trigger_dict
-
-    db_job.args = job.args
-    db_job.kwargs = job.kwargs
-    db_job.max_instances = job.max_instances
-    db_job.coalesce = job.coalesce
-    db_job.misfire_grace_time = job.misfire_grace_time
-
-    db.commit()
-    db.refresh(db_job)
-    logger.info(f"Job '{job_id}' updated successfully.")
-    sync_jobs_from_db() # Sync APScheduler with the new database state
-    return JobConfig.model_validate(db_job)
-
-@app.delete(
-    "/jobs/{job_id}",
-    status_code=status.HTTP_204_NO_CONTENT,
-    tags=["Job Definitions"],
-    responses={
-        status.HTTP_404_NOT_FOUND: {"model": ErrorResponse, "description": "Not Found: Job not found"}
-    }
-)
-def delete_job(job_id: str, db: Session = Depends(get_db)):
-    db_job = db.query(JobDefinition).filter(JobDefinition.id == job_id).first()
-    if db_job is None:
-        logger.warning(f"Attempted to delete non-existent job: {job_id}")
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Job not found"
-        )
-    
-    db.delete(db_job)
-    db.commit()
-    logger.info(f"Job '{job_id}' deleted successfully.")
-    sync_jobs_from_db() # Sync APScheduler with the new database state
-    return
-
-@app.post(
-    "/jobs/bulk/delete",
-    status_code=status.HTTP_200_OK,
-    tags=["Job Definitions"],
-    responses={
-        status.HTTP_404_NOT_FOUND: {"model": ErrorResponse, "description": "One or more jobs not found"}
-    }
-)
-def delete_bulk_jobs(payload: BulkJobUpdate, db: Session = Depends(get_db)):
-    job_ids = payload.job_ids
-    if not job_ids:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="No job IDs provided for deletion."
-        )
-
-    deleted_count = 0
-    not_found_ids = []
-
-    for job_id in job_ids:
-        db_job = db.query(JobDefinition).filter(JobDefinition.id == job_id).first()
-        if db_job:
-            db.delete(db_job)
-            deleted_count += 1
-        else:
-            not_found_ids.append(job_id)
-    
-    if deleted_count > 0:
-        db.commit()
-        logger.info(f"Bulk deleted {deleted_count} jobs.")
-        sync_jobs_from_db()  # Sync once after all deletions
-
-    if not_found_ids:
-        logger.warning(f"Bulk delete: Could not find job IDs: {not_found_ids}")
-        # Even if some are not found, we don't fail the whole request if others succeeded.
-        # The client can be notified of which ones failed.
-        return {"message": f"Deleted {deleted_count} jobs. Not found: {', '.join(not_found_ids)}"}
-
-    return {"message": f"Successfully deleted {deleted_count} jobs."}
-
-@app.get("/api/jobs/{job_id}/history", response_model=List[ProcessExecutionLogInfo], tags=["Job Details"])
-def get_job_execution_history(job_id: str, db: Session = Depends(get_db)):
-    """
-    Retrieves the execution history for a specific job.
-    """
-    try:
-        history = (
-            db.query(ProcessExecutionLog)
-            .filter(ProcessExecutionLog.job_id == job_id)
-            .order_by(ProcessExecutionLog.start_time.desc())
-            .all()
-        )
-        if not history:
-            # Optionally, raise 404 if job_id itself doesn't exist, but for history, returning empty list is fine if no runs yet.
-            # Check if job_id exists at all if strictness is needed.
-            # job_exists = db.query(JobDefinition).filter(JobDefinition.id == job_id).first()
-            # if not job_exists:
-            #     raise HTTPException(
-            #         status_code=status.HTTP_404_NOT_FOUND,
-            #         detail=f"Job '{job_id}' not found."
-            #     )
-            pass # Simply return empty list if no history
-        return history
-    except Exception as e:
-        logger.error(f"Error fetching history for job '{job_id}': {e}", exc_info=True)
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail=f"Failed to fetch execution history for job '{job_id}'."
-        )
-
-
-# --- Scheduler Control Endpoints (APScheduler Instance) ---
-
-@app.get("/scheduler/jobs", response_model=List[JobInfo], tags=["Scheduler Control"])
-def get_scheduled_jobs():
-    """Returns a list of all jobs currently scheduled in the scheduler."""
-    jobs = scheduler.get_jobs()
-    job_infos = []
-    for job in jobs:
-        try:
-            # --- Trigger Conversion ---
-            trigger_dict = {"type": "unknown"}
-            trigger_class_name = job.trigger.__class__.__name__.lower()
-
-            if "cron" in trigger_class_name:
-                trigger_dict["type"] = "cron"
-                for field in job.trigger.fields:
-                    # The string representation of the field is its value.
-                    trigger_dict[field.name] = str(field)
-            elif "interval" in trigger_class_name:
-                trigger_dict["type"] = "interval"
-                td = job.trigger.interval
-                trigger_dict['weeks'] = td.days // 7
-                trigger_dict['days'] = td.days % 7
-                trigger_dict['hours'] = td.seconds // 3600
-                trigger_dict['minutes'] = (td.seconds // 60) % 60
-                trigger_dict['seconds'] = td.seconds % 60
-            elif 'date' in trigger_class_name:
-                # The Pydantic model doesn't currently support DateTrigger, so we skip it.
-                # In a real scenario, you might add a DateTrigger model.
-                logger.warning(f"Skipping job '{job.id}' with unsupported DateTrigger.")
-                continue
-
-            # --- Function Representation ---
-            func_repr = job.func
-            if not isinstance(func_repr, str):
-                func_repr = f"{job.func.__module__}:{job.func.__name__}"
-
-            # --- Assemble JobInfo ---
-            job_info = JobInfo(
-                id=job.id,
-                func=func_repr,
-                trigger=trigger_dict,  # Pydantic will validate and coerce this dict
-                args=list(job.args),
-                kwargs=job.kwargs,
-                max_instances=job.max_instances,
-                coalesce=job.coalesce,
-                misfire_grace_time=job.misfire_grace_time,
-                next_run_time=job.next_run_time
-            )
-            job_infos.append(job_info)
-
-        except Exception as e:
-            logger.error(f"Error processing job '{job.id}' for API response: {e}", exc_info=True)
-
-    return job_infos
-
-@app.post(
-    "/scheduler/jobs/bulk/pause",
-    tags=["Scheduler Control"],
-    responses={
-        status.HTTP_500_INTERNAL_SERVER_ERROR: {"model": ErrorResponse, "description": "Internal Server Error"}
-    }
-)
-def pause_bulk_scheduled_jobs(payload: BulkJobUpdate):
-    job_ids = payload.job_ids
-    paused_ids = []
-    failed_ids = {}
-
-    for job_id in job_ids:
-        try:
-            scheduler.pause_job(job_id)
-            paused_ids.append(job_id)
-        except JobLookupError:
-            failed_ids[job_id] = "Not Found"
-        except Exception as e:
-            failed_ids[job_id] = str(e)
-
-    logger.info(f"Bulk pause request: Paused {len(paused_ids)}. Failed {len(failed_ids)}.")
-    if failed_ids:
-        return {"message": "Partial success", "paused": paused_ids, "failed": failed_ids}
-    
-    return {"message": "All selected jobs paused successfully."}
-
-
-@app.post(
-    "/scheduler/jobs/bulk/resume",
-    tags=["Scheduler Control"],
-    responses={
-        status.HTTP_500_INTERNAL_SERVER_ERROR: {"model": ErrorResponse, "description": "Internal Server Error"}
-    }
-)
-def resume_bulk_scheduled_jobs(payload: BulkJobUpdate):
-    job_ids = payload.job_ids
-    resumed_ids = []
-    failed_ids = {}
-
-    for job_id in job_ids:
-        try:
-            scheduler.resume_job(job_id)
-            resumed_ids.append(job_id)
-        except JobLookupError:
-            failed_ids[job_id] = "Not Found"
-        except Exception as e:
-            failed_ids[job_id] = str(e)
-
-    logger.info(f"Bulk resume request: Resumed {len(resumed_ids)}. Failed {len(failed_ids)}.")
-    if failed_ids:
-        return {"message": "Partial success", "resumed": resumed_ids, "failed": failed_ids}
-        
-    return {"message": "All selected jobs resumed successfully."}
-
-@app.post(
-    "/scheduler/jobs/{job_id}/pause",
-    tags=["Scheduler Control"],
-    responses={
-        status.HTTP_404_NOT_FOUND: {"model": ErrorResponse, "description": "Not Found: Job not found"},
-        status.HTTP_500_INTERNAL_SERVER_ERROR: {"model": ErrorResponse, "description": "Internal Server Error"}
-    }
-)
-def pause_scheduled_job(job_id: str):
-    try:
-        scheduler.pause_job(job_id)
-        logger.info(f"Job '{job_id}' paused successfully.")
-        return {"message": f"Job '{job_id}' paused successfully."}
-    except JobLookupError:
-        logger.warning(f"Attempted to pause non-existent job: '{job_id}'")
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Job '{job_id}' not found."
-        )
-    except Exception as e:
-        logger.error(f"An unexpected error occurred while pausing job '{job_id}': {e}", exc_info=True)
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail=f"An unexpected error occurred while pausing job '{job_id}': {e}"
-        )
-
-@app.post(
-    "/scheduler/jobs/{job_id}/resume",
-    tags=["Scheduler Control"],
-    responses={
-        status.HTTP_404_NOT_FOUND: {"model": ErrorResponse, "description": "Not Found: Job not found"},
-        status.HTTP_500_INTERNAL_SERVER_ERROR: {"model": ErrorResponse, "description": "Internal Server Error"}
-    }
-)
-def resume_scheduled_job(job_id: str):
-    try:
-        scheduler.resume_job(job_id)
-        logger.info(f"Job '{job_id}' resumed successfully.")
-        return {"message": f"Job '{job_id}' resumed successfully."}
-    except JobLookupError:
-        logger.warning(f"Attempted to resume non-existent job: '{job_id}'")
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Job '{job_id}' not found."
-        )
-    except Exception as e:
-        logger.error(f"An unexpected error occurred while resuming job '{job_id}': {e}", exc_info=True)
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail=f"An unexpected error occurred while resuming job '{job_id}': {e}"
-        )
-
-@app.post(
-    "/scheduler/jobs/{job_id}/run",
-    tags=["Scheduler Control"],
-    responses={
-        status.HTTP_404_NOT_FOUND: {"model": ErrorResponse, "description": "Not Found: Job not found"},
-        status.HTTP_500_INTERNAL_SERVER_ERROR: {"model": ErrorResponse, "description": "Internal Server Error"}
-    }
-)
-def run_scheduled_job_immediately(job_id: str):
-    try:
-        # This is the recommended way to trigger a job immediately in APScheduler
-        # It respects max_instances and other job settings
-        scheduler.modify_job(job_id, next_run_time=datetime.now())
-        logger.info(f"Job '{job_id}' scheduled for immediate execution.")
-        return {"message": f"Job '{job_id}' scheduled for immediate execution."}
-    except JobLookupError:
-        logger.warning(f"Attempted to run non-existent job immediately: '{job_id}'")
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Job '{job_id}' not found."
-        )
-    except Exception as e:
-        logger.error(f"An unexpected error occurred while triggering job '{job_id}' for immediate execution: {e}", exc_info=True)
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail=f"An unexpected error occurred while triggering job '{job_id}' for immediate execution: {e}"
-        )
-
-@app.post("/api/scheduler/pause", tags=["Scheduler Control"])
-def pause_global_scheduler():
-    """Pauses the entire scheduler."""
-    scheduler.pause()
-    logger.info("Global scheduler paused.")
-    return {"message": "Scheduler paused successfully."}
-
-@app.post("/api/scheduler/resume", tags=["Scheduler Control"])
-def resume_global_scheduler():
-    """Resumes the entire scheduler."""
-    scheduler.resume()
-    logger.info("Global scheduler resumed.")
-    return {"message": "Scheduler resumed successfully."}
-
-@app.post("/api/scheduler/shutdown", tags=["Scheduler Control"])
-def shutdown_global_scheduler():
-    """Shuts down the entire scheduler. This will stop all jobs and the API."""
-    shutdown_scheduler()
-    logger.info("Global scheduler shut down.")
-    return {"message": "Scheduler shut down successfully. The application will now exit."}
-
-@app.get("/api/jobs_yaml", tags=["Configuration"])
-def get_jobs_yaml_content():
-    """Retrieves the content of the jobs.yaml file."""
-    config_path = "jobs.yaml" # Assuming jobs.yaml is in the root
-    try:
-        with open(config_path, "r", encoding="utf-8") as f:
-            content = f.read()
-        return {"content": content}
-    except FileNotFoundError:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"File not found: {config_path}"
-        )
-    except Exception as e:
-        logger.error(f"Error reading jobs.yaml: {e}", exc_info=True)
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Failed to read jobs.yaml"
-        )
-
-NOTIFICATION_SETTINGS_FILE = "notification_settings.json"
-
-@app.get("/api/settings/notifications", tags=["Settings"])
-def get_notification_settings():
-    """Retrieves notification settings."""
-    if not os.path.exists(NOTIFICATION_SETTINGS_FILE):
-        return {"email_recipients": "", "webhook_url": ""} # Default empty settings
-
-    try:
-        with open(NOTIFICATION_SETTINGS_FILE, "r", encoding="utf-8") as f:
-            settings = json.load(f)
-        return settings
-    except Exception as e:
-        logger.error(f"Error reading notification settings: {e}", exc_info=True)
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Failed to read notification settings."
-        )
-
-@app.post("/api/settings/notifications", tags=["Settings"])
-def update_notification_settings(email_recipients: str = "", webhook_url: str = ""):
-    """Updates notification settings."""
-    settings = {
-        "email_recipients": email_recipients,
-        "webhook_url": webhook_url
-    }
-    try:
-        with open(NOTIFICATION_SETTINGS_FILE, "w", encoding="utf-8") as f:
-            json.dump(settings, f, indent=4)
-        logger.info("Notification settings updated successfully.")
-        return {"message": "Notification settings updated successfully."}
-    except Exception as e:
-        logger.error(f"Error writing notification settings: {e}", exc_info=True)
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Failed to update notification settings."
-        )
-
